refactor: 獨立 StoryProgress 元件
