+++ src/components/StoryContainer.vue
@@ -10,15 +10,12 @@ import {
 } from '../store'
 
 // component
-import ProgressItem from './ProgressItem.vue'
+import StoryProgress from './StoryProgress.vue'
 import StoryImage from './StoryImage.vue'
 
 // story 顯示相關參數
 const isStoryLoaded = computed(() => Object.keys(currentStory).length > 0)
-const itemNum = computed(() => {
-    return mapState.stories.length
-})
-const isCurrentStoryReady = computed(() => mapState.isCurrentStoryReady)
+
 </script>
 
 <template>
@@ -26,22 +23,7 @@ const isCurrentStoryReady = computed(() => mapState.isCurrentStoryReady)
         v-if="!mapState.isLoading && isStoryLoaded"
         class="story-container"
     >
-        <div
-            class="story-progress"
-        >
-            <ProgressItem
-                v-for="(story, index) in mapState.stories"
-                :key="story.id"
-                :current-display-index="mapState.index"
-                :duration="story.duration"
-                :is-current-story-ready="isCurrentStoryReady"
-                :story-index="index"
-                :class="[
-                    'story-progress__item',
-                    mapState.index > index && 'story-progress__item--displayed',
-                ]"
-            />
-        </div>
+        <StoryProgress />
         <button
             class="story-btn prev-btn"
             @click="prevStory"
@@ -56,7 +38,7 @@ const isCurrentStoryReady = computed(() => mapState.isCurrentStoryReady)
         v-else
         class="story-container"
     >
-    <div
+        <div
             class="story-image"
         >
             Loading...
@@ -96,30 +78,6 @@ const isCurrentStoryReady = computed(() => mapState.isCurrentStoryReady)
 
     .next-btn {
         right: 0;
-    }
-
-    .story-progress {
-        position: absolute;
-        top: 12px;
-        left: 50%;
-        transform: translate(-50%, -50%);
-        width: 500px;
-        height: 2px;
-        display: grid;
-        grid-template-columns: v-bind('`repeat(${itemNum}, 1fr)`');
-        gap: 4px;
-        z-index: 1;
-
-        &__item {
-            width: 100%;
-            height: 2px;
-            background: black;
-            position: relative;
-
-            &--displayed {
-                background: white;
-            }
-        }
     }
 }
 </style>
